removed NPM requirement from slack and hand rolled just the chat call
